// src/utils/types.ts
export interface Team {
  team: string;
  name: ChatMessage;
  friendlyFire: number;
  nameTagVisibility: string;
  collisionRule: string;
  color: string;
  prefix: ChatMessage;
  suffix: ChatMessage;
  memberMap: { [name: string]: "" };
  members: string[];

  parseMessage(value: string): ChatMessage;

  add(name: string, value: number): void;

  remove(name: string): void;

  update(
    name: string,
    friendlyFire: boolean,
    nameTagVisibility: string,
    collisionRule: string,
    formatting: number,
    prefix: string,
    suffix: string
  ): void;

  displayName(member: string): ChatMessage;
}